requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
Pillow==10.1.0
python-decouple==3.8 